third_party/blink/web_tests/platform/generic/fast/dom/collection-nameditem-move-between-documents-expected.txt

PASS children.namedItem('child') is null
PASS successfullyParsed is true

TEST COMPLETE
